Create SQL_FirstTestCode
For both queries we need to make sure we filter out keeping cities with >500 order in total, not only breakfast. We need to correct the query 1 having statement and instead of HAVING - add in WHERE a clause that will say (this needs to be added for both queries):

city in (
Select city 
FROM 'bi-2019-test.ad.hoc.orders_jan2021' 
WHERE COUNT(order_id) > 500)

By doing that we make sure we include cities that have more than 500 orders.